Undo FLEX yyleng type change to fix compilation.
The type change of variable yyleng from int to size_t by commit
08eb8c962c3f6d3561217e18419c18fa4b52e670 broke the build on
my system (which is a 64bit system; the commit message indicated that
the change should benefit 64bit?)

+++ src/Headers/globals.h
@@ -35,7 +35,7 @@ extern /*@owned@*/ fileloc g_currentloc;
 /*@-declundef@*/ /* Might not process grammar files */
 extern /*@dependent@*/ /*@open@*/ FILE *yyin;
 extern /*@dependent@*/ /*@open@*/ /*@unused@*/ FILE *yyout;
-extern /*@unused@*/ size_t yyleng;
+extern /*@unused@*/ int yyleng;
 /*@=incondefs@*/ 
 
 extern int yydebug;
